+++ src/components/index.tsx
@@ -1,0 +1,5 @@
+import TD2 from "./common/typography/TD2";
+import Table from "./common/table/Table";
+import Loading from "./common/Loading";
+
+export { TD2, Table, Loading };
